Cutover: move PDS4 logic to PDS4Label, re-parent all leaf labels (issue #297, L-01 step 3)
label.py is now version-agnostic: get_missions/get_observers/get_targets
and the pds_version-gated __init__ blocks move to PDS4Label wholesale;
get_mission_reference_type/get_target_reference_type raise
NotImplementedError on PDSLabel, with PDS4Label providing the
data_to_investigation/data_to_target default. write_label() now reads
self._label_extension/self._eol instead of branching on
setup.pds_version.

All 10 leaf label classes re-parented to PDS3Label or PDS4Label.
ChecksumPDS4Label, DocumentPDS4Label (mission only) and
OrbnumFilePDS4Label (both, conditional on information_model_float) gain
explicit reference-type overrides, replacing the branches they used to
occupy in the deleted class-name dispatch chain. BundlePDS4Label and
InventoryPDS4Label keep their pre-existing overrides unchanged.

Test fallout handled in the same commit, since production and tests
were only ever safe to change together here:
- test_classes_label_pds3_spice_kernel.py and
  test_classes_label_pds4_document.py patched PDSLabel.__init__ by
  qualified name through the leaf module's import; retargeted to
  PDS3Label.__init__/PDS4Label.__init__ to match the new import.
- test_classes_label_pds4_checksum.py, _document.py and
  _orbnum_file.py gain explicit get_mission_reference_type/
  get_target_reference_type assertions for their new overrides,
  mirroring the pattern already used by the bundle/inventory test
  files, so removing the central dispatch-chain test doesn't lose
  coverage.
- test_classes_label.py trimmed to version-agnostic PDSLabel behavior
  only; the pds_version-gated init tests and the three get_missions/
  get_observers/get_targets test classes moved to the new
  test_classes_label_pds4.py. The old class-name dispatch tests are
  replaced by a NotImplementedError contract test on PDSLabel and a
  default-value test on PDS4Label.
- test_classes_label_pds3.py added: thin coverage for PDS3Label, which
  currently adds no logic of its own beyond _label_extension/_eol.

Full suite: pytest tests/npb/ -> 1764 passed, 8 skipped, 1 xfailed,
99% coverage, 100% on every touched label file.
