+++ src/pds/naif_pds4_bundler/classes/label/label.py
@@ -5,7 +5,6 @@ import logging
 import os
 from pathlib import Path
 
-from ...pipeline.runtime import handle_npb_error
 from ...utils import add_carriage_return, compare_files
 
 
@@ -18,14 +17,6 @@ class PDSLabel:
 
     def __init__(self, setup, product) -> None:
         """Constructor."""
-        if setup.pds_version == "4":
-            try:
-                context_products = product.collection.bundle.context_products
-                if not context_products:
-                    raise Exception("No context products from bundle in collection")
-            except BaseException:
-                context_products = product.bundle.context_products
-
         self.name = ''
         self.product = product
         self.setup = setup
@@ -35,64 +26,6 @@ class PDSLabel:
         #
         self.root_dir = setup.root_dir
         self.mission_acronym = setup.mission_acronym
-
-        if setup.pds_version == "4":
-            self.XML_MODEL = setup.xml_model
-            self.SCHEMA_LOCATION = setup.schema_location
-            self.INFORMATION_MODEL_VERSION = setup.information_model
-
-            #
-            # Needs to be built for several Missions.
-            #
-            if hasattr(setup, "secondary_missions"):
-                if len(setup.secondary_missions) == 1:
-                    missions_text = (
-                        f"{setup.mission_name} and {setup.secondary_missions[0]}"
-                    )
-                else:
-                    missions_text = f"{setup.mission_name}, "
-                    for i, sm_name in enumerate(setup.secondary_missions):
-                        if i == len(setup.secondary_missions) - 1:
-                            missions_text += f"and {sm_name}"
-                        else:
-                            missions_text += f"{sm_name}, "
-
-                self.PDS4_MISSION_NAME = f"{missions_text}"
-            else:
-                self.PDS4_MISSION_NAME = f"{setup.mission_name}"
-
-            #
-            # Needs to be built for several observers.
-            #
-            if hasattr(setup, "secondary_observers"):
-                if len(setup.secondary_observers) == 1:
-                    observers_text = (
-                        f"{setup.observer} and {setup.secondary_observers[0]}"
-                    )
-                else:
-                    observers_text = f"{setup.observer}, "
-                    for i, so_name in enumerate(setup.secondary_observers):
-                        if i == len(setup.secondary_observers) - 1:
-                            observers_text += f"and {so_name}"
-                        else:
-                            observers_text += f"{so_name}, "
-
-                self.PDS4_OBSERVER_NAME = f"{observers_text} spacecraft and their"
-            else:
-                self.PDS4_OBSERVER_NAME = f"{setup.observer} spacecraft and its"
-
-            self.END_OF_LINE_PDS4 = "Carriage-Return Line-Feed"
-            if setup.end_of_line == "CRLF":
-                self.END_OF_LINE = "Carriage-Return Line-Feed"
-            elif setup.end_of_line == "LF":
-                self.END_OF_LINE = "Line-Feed"
-            else:
-                handle_npb_error(
-                    "End of Line provided via configuration is not CRLF nor LF.",
-                    setup=self.setup,
-                )
-
-            self.BUNDLE_DESCRIPTION_LID = f"{setup.logical_identifier}:document:spiceds"
 
         if hasattr(self.setup, "creation_date_time"):
             creation_dt = self.setup.creation_date_time
@@ -158,231 +91,36 @@ class PDSLabel:
             self.observers = product.observers
             self.targets = product.targets
 
-        if setup.pds_version == "4":
-            self.MISSIONS = self.get_missions()
-            self.OBSERVERS = self.get_observers()
-            self.TARGETS = self.get_targets()
-
-    def get_missions(self) -> str:
-        """Get the label mission from the context products.
-
-        :return: List of missions to be included in the label
-        """
-        miss = self.missions
-
-        if not isinstance(miss, list):
-            miss = [miss]
-
-        mis_list_for_label = ""
-
-        try:
-            context_products = self.product.collection.bundle.context_products
-        except BaseException:
-            context_products = self.product.bundle.context_products
-
-        eol = self.setup.eol_pds4
-        tab = self.setup.xml_tab
-        for mis in miss:
-            if mis:
-                mis_name = mis
-                mission_lid, mission_type = None, None
-                for product in context_products:
-                    if product["name"][0] == mis_name and (
-                        product["type"][0] == "Mission"
-                        or product["type"][0] == "Other Investigation"
-                    ):
-                        mission_lid = product["lidvid"].split("::")[0]
-                        mission_type = product["type"][0]
-
-                if not mission_lid:
-                    handle_npb_error(
-                        f"LID has not been obtained for mission {mis}.",
-                        setup=self.setup,
-                    )
-
-                mis_list_for_label += (
-                    f"{' ' * 2 * tab}<Investigation_Area>{eol}"
-                    + f"{' ' * 3 * tab}<name>{mis_name}</name>{eol}"
-                    + f"{' ' * 3 * tab}<type>{mission_type}</type>{eol}"
-                    + f"{' ' * 3 * tab}<Internal_Reference>{eol}"
-                    + f"{' ' * 4 * tab}<lid_reference>{mission_lid}"
-                    f"</lid_reference>{eol}" + f"{' ' * 4 * tab}<reference_type>"
-                    f"{self.get_mission_reference_type()}"
-                    f"</reference_type>{eol}"
-                    + f"{' ' * 3 * tab}</Internal_Reference>{eol}"
-                    + f"{' ' * 2 * tab}</Investigation_Area>{eol}"
-                )
-        if not mis_list_for_label:
-            handle_npb_error(
-                f"{self.product.name} missions not defined.", setup=self.setup
-            )
-        mis_list_for_label = mis_list_for_label.rstrip() + eol
-
-        return mis_list_for_label
-
     def get_mission_reference_type(self):
         """Get the mission reference type.
 
-        :return: Mission_Reference_Type value for PDS4 label
-        :rtype: str
+        Version-agnostic labels have no notion of a mission reference type;
+        only PDS4 labels do. Subclasses that need one must override this.
+
+        :raises NotImplementedError: always, on the base class
         """
-        if self.__class__.__name__ == "ChecksumPDS4Label":
-            ref_type = "ancillary_to_investigation"
-        elif self.__class__.__name__ == "BundlePDS4Label":
-            ref_type = "bundle_to_investigation"
-        elif self.__class__.__name__ == "DocumentPDS4Label":
-            ref_type = "document_to_investigation"
-        elif self.__class__.__name__ == "InventoryPDS4Label":
-            ref_type = "collection_to_investigation"
-        elif self.__class__.__name__ == "OrbnumFilePDS4Label":
-            if self.setup.information_model_float >= 1014000000.0:
-                ref_type = "ancillary_to_investigation"
-            else:
-                ref_type = "data_to_investigation"
-        else:
-            ref_type = "data_to_investigation"
-
-        return ref_type
-
-    def get_observers(self) -> str:
-        """Get the label observers from the context products.
-
-        :return: List of Observers to be included in the label
-        """
-        obs = self.observers
-        if not isinstance(obs, list):
-            obs = [obs]
-
-        obs_list_for_label = ""
-
-        try:
-            context_products = self.product.collection.bundle.context_products
-        except BaseException:
-            context_products = self.product.bundle.context_products
-
-        eol = self.setup.eol_pds4
-        tab = self.setup.xml_tab
-
-        for ob in obs:
-            if ob:
-                ob_lid, ob_type = None, None
-                ob_name = ob.split(",")[0]
-                for product in context_products:
-                    if product["name"][0] == ob_name and (
-                        product["type"][0] == "Spacecraft"
-                        or product["type"][0] == "Rover"
-                        or product["type"][0] == "Lander"
-                        or product["type"][0] == "Host"
-                    ):
-                        ob_lid = product["lidvid"].split("::")[0]
-                        ob_type = product["type"][0]
-
-                if not ob_lid:
-                    handle_npb_error(
-                        f"LID has not been obtained for observer {ob}.",
-                        setup=self.setup,
-                    )
-
-                obs_list_for_label += (
-                    f"{' ' * 3 * tab}<Observing_System_Component>{eol}"
-                    + f"{' ' * (3+1) * tab}<name>{ob_name}</name>{eol}"
-                    + f"{' ' * (3+1) * tab}<type>{ob_type}</type>{eol}"
-                    + f"{' ' * (3+1) * tab}<Internal_Reference>{eol}"
-                    + f"{' ' * (3 + 2) * tab}<lid_reference>{ob_lid}"
-                    f"</lid_reference>{eol}"
-                    + f"{' ' * (3 + 2) * tab}<reference_type>is_instrument_host"
-                    f"</reference_type>{eol}"
-                    + f"{' ' * (3+1) * tab}</Internal_Reference>{eol}"
-                    + f"{' ' * 3 * tab}</Observing_System_Component>{eol}"
-                )
-
-        if not obs_list_for_label:
-            handle_npb_error(
-                f"{self.product.name} observers not defined.", setup=self.setup
-            )
-        obs_list_for_label = obs_list_for_label.rstrip() + eol
-
-        return obs_list_for_label
-
-    def get_targets(self) -> str:
-        """Get the label targets from the context products.
-
-        :return: List of Targets to be included in the label
-        """
-        tars = self.targets
-        if not isinstance(tars, list):
-            tars = [tars]
-
-        tar_list_for_label = ""
-
-        try:
-            context_products = self.product.collection.bundle.context_products
-        except BaseException:
-            context_products = self.product.bundle.context_products
-
-        eol = self.setup.eol_pds4
-        tab = self.setup.xml_tab
-
-        for tar in tars:
-            if tar:
-                target_name = tar
-                target_lid, target_type = None, None
-                for product in context_products:
-                    if product["name"][0].upper() == target_name.upper():
-                        target_lid = product["lidvid"].split("::")[0]
-                        target_type = product["type"][0].capitalize()
-
-                tar_list_for_label += (
-                    f"{' ' * 2 * tab}<Target_Identification>{eol}"
-                    + f"{' ' * 3 * tab}<name>{target_name}</name>{eol}"
-                    + f"{' ' * 3 * tab}<type>{target_type}</type>{eol}"
-                    + f"{' ' * 3 * tab}<Internal_Reference>{eol}"
-                    + f"{' ' * 4 * tab}<lid_reference>{target_lid}"
-                    f"</lid_reference>{eol}" + f"{' ' * 4 * tab}<reference_type>"
-                    f"{self.get_target_reference_type()}"
-                    f"</reference_type>{eol}"
-                    + f"{' ' * 3 * tab}</Internal_Reference>{eol}"
-                    + f"{' ' * 2 * tab}</Target_Identification>{eol}"
-                )
-
-        if not tar_list_for_label:
-            handle_npb_error(f"{self.product.name} targets not defined.", setup=self.setup)
-        tar_list_for_label = tar_list_for_label.rstrip() + eol
-
-        return tar_list_for_label
+        raise NotImplementedError(
+            f"{type(self).__name__} does not define a mission reference type."
+        )
 
     def get_target_reference_type(self):
         """Get the target reference type.
 
-        :return: Target_Reference_Type value for PDS4 label
-        :rtype: str
-        """
-        if self.__class__.__name__ == "ChecksumPDS4Label":
-            ref_type = "ancillary_to_target"
-        elif self.__class__.__name__ == "BundlePDS4Label":
-            ref_type = "bundle_to_target"
-        elif self.__class__.__name__ == "InventoryPDS4Label":
-            ref_type = "collection_to_target"
-        elif self.__class__.__name__ == "OrbnumFilePDS4Label":
-            if self.setup.information_model_float >= 1014000000.0:
-                ref_type = "ancillary_to_target"
-            else:
-                ref_type = "data_to_target"
-        else:
-            ref_type = "data_to_target"
+        Version-agnostic labels have no notion of a target reference type;
+        only PDS4 labels do. Subclasses that need one must override this.
 
-        return ref_type
+        :raises NotImplementedError: always, on the base class
+        """
+        raise NotImplementedError(
+            f"{type(self).__name__} does not define a target reference type."
+        )
 
     def write_label(self):
         """Write the Label."""
         label_dictionary = vars(self)
 
-        if self.setup.pds_version == "4":
-            label_extension = ".xml"
-            eol = self.setup.eol_pds4
-        else:
-            label_extension = ".lbl"
-            eol = self.setup.eol_pds3
+        label_extension = self._label_extension
+        eol = self._eol
 
         if label_extension not in self.product.path:
             label_name = (
