+++ src/pds/naif_pds4_bundler/classes/label/pds4_inventory.py
@@ -3,10 +3,10 @@ files.
 """
 from pathlib import Path
 
-from .label import PDSLabel
+from .pds4_label import PDS4Label
 
 
-class InventoryPDS4Label(PDSLabel):
+class InventoryPDS4Label(PDS4Label):
     """Class to generate a PDS4 Collection Inventory Label.
 
     :param setup:      NPB execution Setup object
