+++ tests/npb/test_classes_label_pds4.py
@@ -1,0 +1,477 @@
+"""
+Tests for PDS4Label class.
+
+PDS4Label carries everything that used to live behind
+``if setup.pds_version == "4":`` guards inside PDSLabel.__init__, plus the
+three PDS4-only context-product methods (get_missions/get_observers/
+get_targets) and the default reference-type values. See
+test_classes_label.py for the version-agnostic PDSLabel coverage and
+test_classes_label_pds3.py for the (currently much thinner) PDS3Label
+coverage.
+"""
+
+from unittest.mock import MagicMock
+
+import pytest
+
+from pds.naif_pds4_bundler.classes.label.pds4_label import PDS4Label
+
+# Patch target — resolved to where the name is looked up inside pds4_label.py
+_PATCH_HANDLE_ERROR = "pds.naif_pds4_bundler.classes.label.pds4_label.handle_npb_error"
+
+
+# ---------------------------------------------------------------------------
+# Shared builder helpers
+# ---------------------------------------------------------------------------
+
+def _make_context_products():
+    """Return a minimal list of context product dicts used across tests."""
+    return [
+        {
+            "name": ["TestMission"],
+            "type": ["Mission"],
+            "lidvid": "urn:nasa:pds:testmission::1.0",
+        },
+        {
+            "name": ["TestObserver"],
+            "type": ["Spacecraft"],
+            "lidvid": "urn:nasa:pds:testobserver::1.0",
+        },
+        {
+            "name": ["TestTarget"],
+            "type": ["planet"],
+            "lidvid": "urn:nasa:pds:testtarget::1.0",
+        },
+    ]
+
+
+def _make_bundle(context_products=None):
+    bundle = MagicMock()
+    bundle.context_products = context_products or _make_context_products()
+    return bundle
+
+
+def _make_collection(bundle=None):
+    collection = MagicMock()
+    collection.bundle = bundle or _make_bundle()
+    collection.name = "spice_kernels"
+    return collection
+
+
+def _make_product(collection=None, bundle=None):
+    """Return a mock product that supplies all attributes PDS4Label touches."""
+    product = MagicMock()
+    product.collection = collection or _make_collection()
+    product.bundle = bundle or _make_bundle()
+    product.creation_time = "2024-01-01T00:00:00"
+    product.creation_date = "2024-01-01"
+    product.size = 1024
+    product.checksum = "abc123"
+    product.missions = ["TestMission"]
+    product.observers = ["TestObserver"]
+    product.targets = ["TestTarget"]
+    product.name = "test_kernel.bc"
+    product.extension = "bc"
+    product.path = "/staging/test_kernel.bc"
+    product.record_bytes = 80
+    return product
+
+
+def _make_setup_pds4(**kwargs):
+    """Return a mock setup object configured for PDS4."""
+    setup = MagicMock()
+    setup.pds_version = "4"
+    setup.root_dir = "/root"
+    setup.mission_acronym = "test"
+    setup.xml_model = "model"
+    setup.schema_location = "schema"
+    setup.information_model = ".".join(["1", "14", "0", "0"])
+    setup.information_model_float = 1014000000.0
+    setup.mission_name = "TestMission"
+    setup.observer = "TestObserver"
+    setup.target = "TestTarget"
+    setup.end_of_line = "CRLF"
+    setup.logical_identifier = "urn:nasa:pds:testmission"
+    setup.eol_pds4 = "\r\n"
+    setup.xml_tab = 2
+    setup.staging_directory = "/staging"
+    setup.working_directory = "/work"
+    setup.bundle_directory = "/bundle"
+    setup.diff = False
+    setup.args.silent = False
+    setup.args.verbose = False
+
+    # Ensure secondary_* attributes are NOT present by default
+    del setup.secondary_missions
+    del setup.secondary_observers
+    del setup.secondary_targets
+    del setup.creation_date_time
+
+    for k, v in kwargs.items():
+        setattr(setup, k, v)
+    return setup
+
+
+# ---------------------------------------------------------------------------
+# Shared fixtures
+# ---------------------------------------------------------------------------
+
+@pytest.fixture
+def setup_pds4():
+    return _make_setup_pds4()
+
+
+@pytest.fixture
+def product():
+    return _make_product()
+
+
+# ===========================================================================
+# PDS4Label.__init__
+# ===========================================================================
+
+class TestPDS4LabelInit:
+    """Covers PDS4Label.__init__ – all branches moved from PDSLabel."""
+
+    @staticmethod
+    @pytest.fixture
+    def mock_class_methods(mocker):
+        # Mocks all three methods at once
+        mocker.patch('pds.naif_pds4_bundler.classes.label.pds4_label.PDS4Label.get_missions',
+                     return_value='MockedMission')
+        mocker.patch('pds.naif_pds4_bundler.classes.label.pds4_label.PDS4Label.get_observers',
+                     return_value='MockedObserver')
+        mocker.patch('pds.naif_pds4_bundler.classes.label.pds4_label.PDS4Label.get_targets',
+                     return_value='MockedTarget')
+
+    def test_pds4_context_from_collection_bundle(self, setup_pds4, product):
+        """context products from collection.bundle"""
+        label = PDS4Label(setup_pds4, product)
+        assert label.setup is setup_pds4
+        assert label.product is product
+        assert label.name == ""
+
+    def test_pds4_context_falls_back_to_product_bundle(self, mock_class_methods, setup_pds4, product):
+        """context products from product.bundle
+           (collection.bundle.context_products is falsy => raises Exception)
+        """
+        product.collection.bundle.context_products = []
+        label = PDS4Label(setup_pds4, product)
+        assert label is not None
+
+    def test_pds4_context_attribute_error_falls_back(self, setup_pds4, product):
+        """collection.bundle raises AttributeError entirely"""
+        product.collection = MagicMock(spec=[])
+        label = PDS4Label(setup_pds4, product)
+        assert label is not None
+
+    def test_pds4_single_mission_name(self, mock_class_methods, setup_pds4, product):
+        label = PDS4Label(setup_pds4, product)
+        assert label.PDS4_MISSION_NAME == "TestMission"
+
+    @pytest.mark.parametrize("secondary_missions, expected_name",[
+        (["OtherMission"], "TestMission and OtherMission"),
+        (["MissionB", "MissionC"], "TestMission, MissionB, and MissionC"),
+    ])
+    def test_pds4_multiple_mission_name(self, mock_class_methods, setup_pds4, product, secondary_missions, expected_name):
+        setup_pds4.secondary_missions = secondary_missions
+        label = PDS4Label(setup_pds4, product)
+        assert label.PDS4_MISSION_NAME == expected_name
+
+    def test_pds4_single_observer_name(self, mock_class_methods, setup_pds4, product):
+        label = PDS4Label(setup_pds4, product)
+        assert label.PDS4_OBSERVER_NAME == "TestObserver spacecraft and its"
+
+    @pytest.mark.parametrize("secondary_observer, expected_name",[
+        (["OtherObserver"], "TestObserver and OtherObserver spacecraft and their"),
+        (["ObsB", "ObsC"], "TestObserver, ObsB, and ObsC spacecraft and their"),
+    ])
+    def test_pds4_multiple_observer_name(self, mock_class_methods, setup_pds4, product, secondary_observer, expected_name):
+        setup_pds4.secondary_observers = secondary_observer
+        label = PDS4Label(setup_pds4, product)
+        assert label.PDS4_OBSERVER_NAME == expected_name
+
+    @pytest.mark.parametrize("eol, expected_eol_name",[
+        ('LF', 'Line-Feed'),
+        ('CRLF', 'Carriage-Return Line-Feed')
+    ])
+    def test_pds4_end_of_line(self, product, eol, expected_eol_name):
+        setup = _make_setup_pds4(end_of_line=eol)
+        label = PDS4Label(setup, product)
+        assert label.END_OF_LINE == expected_eol_name
+
+    def test_pds4_end_of_line_invalid(self, product):
+        """end_of_line is invalid (neither CRLF nor LF)"""
+        setup = _make_setup_pds4(end_of_line="CR")
+        with pytest.raises(RuntimeError, match=r'End of Line provided via configuration '
+                                               r'is not CRLF nor LF\.'):
+            PDS4Label(setup, product)
+
+    # TODO: The following two test cases demonstrate an issue:
+    #       If they are not a list, the code to set up PDS4_MISSION_NAME
+    #       and PDS4_OBSERVER_NAME, results in a list of letters.
+    def test_pds4_secondary_missions_non_list_wrapped(self, mock_class_methods, setup_pds4, product):
+        setup_pds4.secondary_missions = "SingleMission"
+        label = PDS4Label(setup_pds4, product)
+        assert label.missions == ['TestMission', 'SingleMission']
+        # TODO: This is a bug!!!!
+        assert 'TestMission, S, i, n, g, l, e, M, i, s, s, i, o, and n' == label.PDS4_MISSION_NAME
+
+    def test_pds4_secondary_observers_non_list_wrapped(self, mock_class_methods, setup_pds4, product):
+        setup_pds4.secondary_observers = 'SingleObserver'
+        label = PDS4Label(setup_pds4, product)
+        assert label.observers == ['TestObserver', 'SingleObserver']
+        # TODO: This is a bug!!!
+        assert 'TestObserver, S, i, n, g, l, e, O, b, s, e, r, v, e, and r spacecraft and their' == label.PDS4_OBSERVER_NAME
+
+    def test_pds4_sets_missions_targets_observers(self, mock_class_methods, setup_pds4, product):
+        setup_pds4.secondary_missions = ["MissionB", "MissionC"]
+        setup_pds4.secondary_observers = ["ObsB", "ObsC"]
+        setup_pds4.secondary_targets = ["TargetB", "TargetC"]
+
+        # Since the mock is called, the values of the MISSIONS,
+        # OBSERVERS and TARGETS should be the ones provided as return_value
+        # in the mock.
+        label = PDS4Label(setup_pds4, product)
+        assert label.MISSIONS == 'MockedMission'
+        assert label.OBSERVERS == 'MockedObserver'
+        assert label.TARGETS == 'MockedTarget'
+
+
+# ===========================================================================
+# PDS4Label.get_missions
+# ===========================================================================
+
+class TestPDS4LabelGetMissions:
+    """Covers PDS4Label.get_missions – all branches."""
+
+    @pytest.fixture
+    def label_for(self):
+        """Factory fixture: returns a callable that builds a bare label."""
+        def _build(missions, context_products=None, fallback_to_bundle=False):
+            setup = _make_setup_pds4()
+            product = _make_product()
+            ctx = context_products or _make_context_products()
+            if fallback_to_bundle:
+                product.collection = MagicMock(spec=[])  # no .bundle
+                product.bundle.context_products = ctx
+            else:
+                product.collection.bundle.context_products = ctx
+            label = PDS4Label.__new__(PDS4Label)
+            label.setup = setup
+            label.product = product
+            label.missions = missions if isinstance(missions, list) else [missions]
+            return label
+        return _build
+
+    @pytest.mark.parametrize('missions, expected', [
+        (['TestMission'], '    <Investigation_Area>\r\n'
+                          '      <name>TestMission</name>\r\n'
+                          '      <type>Mission</type>\r\n'
+                          '      <Internal_Reference>\r\n'
+                          '        <lid_reference>urn:nasa:pds:testmission</lid_reference>\r\n'
+                          '        <reference_type>data_to_investigation</reference_type>\r\n'
+                          '      </Internal_Reference>\r\n'
+                          '    </Investigation_Area>\r\n'),
+        ('TestMission', '    <Investigation_Area>\r\n'
+                        '      <name>TestMission</name>\r\n'
+                        '      <type>Mission</type>\r\n'
+                        '      <Internal_Reference>\r\n'
+                        '        <lid_reference>urn:nasa:pds:testmission</lid_reference>\r\n'
+                        '        <reference_type>data_to_investigation</reference_type>\r\n'
+                        '      </Internal_Reference>\r\n'
+                        '    </Investigation_Area>\r\n')
+    ])
+    def test_mission(self, label_for, missions, expected):
+        label = label_for(missions)
+        label.missions = missions
+        result = label.get_missions()
+        assert result == expected
+
+    def test_context_falls_back_to_product_bundle(self, label_for):
+        result = label_for(["TestMission"], fallback_to_bundle=True).get_missions()
+        assert "TestMission" in result
+
+    def test_empty_mission_name_skipped_calls_error(self, label_for, mocker):
+        """A falsy mission entry (empty string) must be skipped."""
+        mock_err = mocker.patch(_PATCH_HANDLE_ERROR)
+        label_for([""]).get_missions()
+        mock_err.assert_called()
+
+    def test_other_investigation_type_accepted(self, label_for):
+        ctx = [
+            {
+                "name": ["TestMission"],
+                "type": ["Other Investigation"],
+                "lidvid": "urn:nasa:pds:testmission::1.0",
+            }
+        ]
+        result = label_for(["TestMission"], context_products=ctx).get_missions()
+        assert result == ('    <Investigation_Area>\r\n'
+                          '      <name>TestMission</name>\r\n'
+                          '      <type>Other Investigation</type>\r\n'
+                          '      <Internal_Reference>\r\n'
+                          '        <lid_reference>urn:nasa:pds:testmission</lid_reference>\r\n'
+                          '        <reference_type>data_to_investigation</reference_type>\r\n'
+                          '      </Internal_Reference>\r\n'
+                          '    </Investigation_Area>\r\n')
+
+    def test_no_lid_found_calls_handle_npb_error(self, label_for, mocker):
+        """No context product matches → mission_lid never set → handle_npb_error."""
+        ctx = [
+            {
+                "name": ["Different"],
+                "type": ["Mission"],
+                "lidvid": "urn:nasa:pds:different::1.0",
+            }
+        ]
+        mock_err = mocker.patch(_PATCH_HANDLE_ERROR)
+        label_for(["TestMission"], context_products=ctx).get_missions()
+        mock_err.assert_called()
+
+
+# ===========================================================================
+# PDS4Label.get_observers
+# ===========================================================================
+
+class TestPDS4LabelGetObservers:
+    """Covers PDS4Label.get_observers – all branches."""
+
+    @pytest.fixture
+    def label_for(self):
+        """Factory fixture: builds a bare label with given observers."""
+        def _build(observers, context_products=None, fallback_to_bundle=False):
+            setup = _make_setup_pds4()
+            product = _make_product()
+            ctx = context_products or _make_context_products()
+            if fallback_to_bundle:
+                product.collection = MagicMock(spec=[])
+                product.bundle.context_products = ctx
+            else:
+                product.collection.bundle.context_products = ctx
+            label = PDS4Label.__new__(PDS4Label)
+            label.setup = setup
+            label.product = product
+            label.observers = observers if isinstance(observers, list) else [observers]
+            return label
+        return _build
+
+    def test_single_observer_spacecraft(self, label_for):
+        result = label_for(["TestObserver"]).get_observers()
+        assert "TestObserver" in result
+        assert "Observing_System_Component" in result
+
+    def test_observers_not_a_list_wrapped(self, label_for):
+        label = label_for("TestObserver")
+        label.observers = "TestObserver"  # scalar
+        assert "TestObserver" in label.get_observers()
+
+    def test_context_fallback_to_product_bundle(self, label_for):
+        result = label_for(["TestObserver"], fallback_to_bundle=True).get_observers()
+        assert "TestObserver" in result
+
+    def test_rover_type_accepted(self, label_for):
+        ctx = [{"name": ["TestObserver"], "type": ["Rover"], "lidvid": "urn:x::1.0"}]
+        assert "Rover" in label_for(["TestObserver"], ctx).get_observers()
+
+    def test_lander_type_accepted(self, label_for):
+        ctx = [{"name": ["TestObserver"], "type": ["Lander"], "lidvid": "urn:x::1.0"}]
+        assert "Lander" in label_for(["TestObserver"], ctx).get_observers()
+
+    def test_host_type_accepted(self, label_for):
+        ctx = [{"name": ["TestObserver"], "type": ["Host"], "lidvid": "urn:x::1.0"}]
+        assert "Host" in label_for(["TestObserver"], ctx).get_observers()
+
+    def test_comma_in_observer_name_strips_suffix(self, label_for):
+        """Observer names like 'Name, suffix' → only 'Name' used for look-up."""
+        ctx = [{"name": ["TestObserver"], "type": ["Spacecraft"], "lidvid": "urn:x::1.0"}]
+        assert "TestObserver" in label_for(["TestObserver, extra"], ctx).get_observers()
+
+    def test_empty_observer_skipped_calls_error(self, label_for, mocker):
+        mock_err = mocker.patch(_PATCH_HANDLE_ERROR)
+        label_for([""]).get_observers()
+        mock_err.assert_called()
+
+    def test_no_lid_calls_handle_npb_error(self, label_for, mocker):
+        ctx = [{"name": ["NoMatch"], "type": ["Spacecraft"], "lidvid": "urn:x::1.0"}]
+        mock_err = mocker.patch(_PATCH_HANDLE_ERROR)
+        label_for(["TestObserver"], ctx).get_observers()
+        mock_err.assert_called()
+
+    def test_trailing_eol_appended(self, label_for):
+        assert label_for(["TestObserver"]).get_observers().endswith("\r\n")
+
+
+# ===========================================================================
+# PDS4Label.get_targets
+# ===========================================================================
+
+class TestPDS4LabelGetTargets:
+    """Covers PDS4Label.get_targets – all branches."""
+
+    @pytest.fixture
+    def label_for(self):
+        """Factory fixture: builds a bare label with given targets."""
+        def _build(targets, context_products=None, fallback_to_bundle=False):
+            setup = _make_setup_pds4()
+            product = _make_product()
+            ctx = context_products or _make_context_products()
+            if fallback_to_bundle:
+                product.collection = MagicMock(spec=[])
+                product.bundle.context_products = ctx
+            else:
+                product.collection.bundle.context_products = ctx
+            label = PDS4Label.__new__(PDS4Label)
+            label.setup = setup
+            label.product = product
+            label.targets = targets if isinstance(targets, list) else [targets]
+            return label
+        return _build
+
+    def test_single_target(self, label_for):
+        result = label_for(["TestTarget"]).get_targets()
+        assert "TestTarget" in result
+        assert "Target_Identification" in result
+
+    def test_targets_not_a_list_wrapped(self, label_for):
+        label = label_for("TestTarget")
+        label.targets = "TestTarget"  # scalar
+        assert "TestTarget" in label.get_targets()
+
+    def test_context_fallback_to_product_bundle(self, label_for):
+        result = label_for(["TestTarget"], fallback_to_bundle=True).get_targets()
+        assert "TestTarget" in result
+
+    def test_case_insensitive_name_match(self, label_for):
+        ctx = [{"name": ["TESTTARGET"], "type": ["planet"], "lidvid": "urn:x::1.0"}]
+        assert "testtarget" in label_for(["testtarget"], ctx).get_targets()
+
+    def test_empty_target_skipped_calls_error(self, label_for, mocker):
+        mock_err = mocker.patch(_PATCH_HANDLE_ERROR)
+        label_for([""]).get_targets()
+        mock_err.assert_called()
+
+    def test_trailing_eol_appended(self, label_for):
+        assert label_for(["TestTarget"]).get_targets().endswith("\r\n")
+
+
+# ===========================================================================
+# PDS4Label.get_mission_reference_type / get_target_reference_type
+# ===========================================================================
+
+class TestPDS4LabelDefaultReferenceTypes:
+    """Covers PDS4Label's own default reference types.
+
+    Per-leaf overrides (Checksum/Bundle/Document/Inventory/OrbnumFile) are
+    covered in each leaf class's own test file; SpiceKernelPDS4Label and
+    MetaKernelPDS4Label have no override and so are covered by inheriting
+    this same default, exercised here directly on PDS4Label.
+    """
+
+    def test_default_mission_reference_type(self):
+        label = object.__new__(PDS4Label)
+        assert label.get_mission_reference_type() == "data_to_investigation"
+
+    def test_default_target_reference_type(self):
+        label = object.__new__(PDS4Label)
+        assert label.get_target_reference_type() == "data_to_target"
